Refactor tests for moved myservice module
