+++ cfme/tests/services/test_add_remove_vm_to_service.py
@@ -1,18 +1,16 @@
 # -*- coding: utf-8 -*-
 import fauxfactory
 import pytest
-
-from cfme.common.provider import cleanup_vm
-from cfme.services.catalogs.service_catalogs import ServiceCatalogs
-from cfme.services import requests
-from cfme.services.catalogs.myservice import MyService
-from cfme.automate.simulation import simulate
-from cfme.automate.explorer import Domain, Namespace, Class, Method
 from cfme import test_requirements
+from cfme.automate.explorer import Class, Domain, Method, Namespace
+from cfme.automate.simulation import simulate
+from cfme.common.provider import cleanup_vm
+from cfme.services import requests
+from cfme.services.catalogs.service_catalogs import ServiceCatalogs
+from cfme.services.myservice import MyService
 from utils import testgen
 from utils.log import logger
 from utils.wait import wait_for
-
 
 pytestmark = [
     test_requirements.service,
